app/requirements.txt
aiogram>=3.4.1,<4.0.0
aiosqlite>=0.19.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
